+++ src/main/java/net/easecation/ghosty/recording/SkinlessPlayerRecord.java
@@ -79,7 +79,7 @@ public class SkinlessPlayerRecord implements PlayerRecord {
 
         private RecordPair(BinaryStream stream) {
             try {
-                this.tick = stream.getUnsignedVarInt();
+                this.tick = (int) stream.getUnsignedVarInt();
                 this.updated = Updated.fromBinaryStream(stream);
             } catch (Exception e) {
                 Server.getInstance().getLogger().logException(e);
